increase width of base_count to cope with T. aestivum

+++ sql/genome_info.sql
@@ -59,7 +59,7 @@ CREATE TABLE `genome` (
   `assembly_id` varchar(16) DEFAULT NULL,
   `assembly_name` varchar(200) NOT NULL,
   `assembly_level` varchar(50) NOT NULL,
-  `base_count` int(10) unsigned NOT NULL,
+  `base_count` bigint(20) unsigned NOT NULL,
   `genebuild` varchar(64) NOT NULL,
   `dbname` varchar(64) NOT NULL,
   `species_id` int(10) unsigned NOT NULL,
